server_proxy/interview_proxy/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
google-cloud-speech
